+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 
+- Add LSP document highlighting support.
 - Add LSP completion for fully-qualified type references.
 
 ## [v1.63.0] - 2026-01-06